Widen a supplied fact count to 64 bits
The ds64 sampleCount field is 64-bit, so Fact::Samples(u32) could not
state the count for a file long enough to need RF64, which for an
unmodelled format is the only way it gets a count at all. Writing an
oversized count to a 32-bit RIFF fact field is now an error rather than
a truncation.

+++ examples/write_raw.rs
@@ -60,7 +60,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     // the file carries no trustworthy frame count at all. A supplied count is
     // written as is and left alone by finalize.
     let mut writer = WavWriter::builder(fmt)?
-        .fact(Fact::Samples(frames))
+        .fact(Fact::Samples(frames as u64))
         .open(std::fs::File::create(&path)?)?;
 
     let sine: Vec<i16> = (0..frames)
